temp/home_credit_pipeline_backup/pyproject.toml
[project]
name = "home-credit-pipeline"
version = "0.1.0"
description = "Home Credit Default Risk regression pipeline"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "pandas>=2.0",
    "numpy>=1.25",
    "scikit-learn>=1.3",
    "lightgbm>=4.0",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "seaborn>=0.13.2",
    "optuna>=4.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "black>=24.0",
    "mypy>=1.10"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/home_credit --cov-report=term-missing"
